encapsulate attributes and provide extensions to access ElementEntity.affiliation
