--- game/engine/src/main/kotlin/de/gleex/pltcmd/game/engine/entities/EntityTypes.kt
@@ -1,7 +1,0 @@
-package de.gleex.pltcmd.game.engine.entities
-
-import de.gleex.pltcmd.game.engine.entities.types.Movable
-import org.hexworks.amethyst.api.base.BaseEntityType
-
-/** Represents an element in an army. */
-object ElementType : BaseEntityType("element", "A movable element."), Movable
