+++ game/engine/src/main/kotlin/de/gleex/pltcmd/game/engine/entities/types/Positionable.kt
@@ -1,7 +1,7 @@
 package de.gleex.pltcmd.game.engine.entities.types
 
 import de.gleex.pltcmd.game.engine.attributes.PositionAttribute
-import de.gleex.pltcmd.game.engine.extensions.PositionableEntity
+import de.gleex.pltcmd.game.engine.extensions.GameEntity
 import de.gleex.pltcmd.game.engine.extensions.getAttribute
 import de.gleex.pltcmd.model.world.coordinate.Coordinate
 import org.hexworks.amethyst.api.entity.EntityType
@@ -15,6 +15,7 @@ import org.hexworks.cobalt.databinding.api.value.ObservableValue
 
 /** Type marker for entities that have the PositionAttribute */
 interface Positionable : EntityType
+typealias PositionableEntity = GameEntity<Positionable>
 
 /** Access to the [Property] of the [PositionAttribute] of a [PositionableEntity] */
 var PositionableEntity.position: ObservableValue<Coordinate>
