+++ game/ui/src/main/kotlin/de/gleex/pltcmd/game/ui/GameView.kt
@@ -1,7 +1,7 @@
 package de.gleex.pltcmd.game.ui
 
 import de.gleex.pltcmd.game.engine.GameContext
-import de.gleex.pltcmd.game.engine.entities.ElementType
+import de.gleex.pltcmd.game.engine.entities.types.ElementType
 import de.gleex.pltcmd.game.options.GameOptions
 import de.gleex.pltcmd.game.options.UiOptions
 import de.gleex.pltcmd.game.ui.entities.GameBlock
